+++ Runtime/Scripts/GlobalTools/Generic/ObjectHelpers_PrimitiveTypes.cs
@@ -2,6 +2,9 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+using System;
+using System.Globalization;
+
 namespace PolytopeSolutions.Toolset.GlobalTools.Generic {
 	public static partial class ObjectHelpers {
 		public static T GetItem<T>(this List<T> list, int index){
@@ -9,5 +12,9 @@ namespace PolytopeSolutions.Toolset.GlobalTools.Generic {
 			                    0, list.Count);
 			return list[index];
 		}
+
+		public static float InvariantConvertToSingle(object item) {
+			return Convert.ToSingle(item, CultureInfo.InvariantCulture.NumberFormat);
+        }
     }
 }